requirements.txt
google-generativeai>=0.8.0
gspread>=6.1.0
google-auth>=2.29.0
requests>=2.31.0
